Deleting sub-item3d objects previously didn't return the "object" to the mesh tree where it was originally split from.
While there are no current examples which do this, it is entirely plausible that a user may want to achieve something like this in the future, and so the capability should be there to return objects which have been "pruned" from the tree/graph of a mesh back to their original position.

This commit achieves this... though requires extensive testing to confirm its stability.