Fix section issue with cube and add tests.
Created unit tests for QGLCube.  These are designed to be implementation
agnostic - they look for the vertex values being present anywhere in
the finalized geometry.

QGLCube does not work when it creates sections internally - the user must specify
sections explicitly.  They may want a smooth cube, or to add a cube and some
other primitives to the same section.