For now displaylist requires newSection() before adding.
QGLCube has been modified to take out the newSection() because otherwise
you can't add multiple primitives to a section that contains a cube, and
you can't create a cube with anything other than QGL::Faceted.

In future might add an automatic initial section to displaylist so
that the requirement to remember to call newSection() goes away.