Fix balancing.
Track test failures and fix balancing bugs.  Add some debug to help with
this and also a few other minor fixes and polishes.

Currently the balancing is sub-optimal - it just looks for chains
where the parent and grandparent both have no branches.  This does
reduce height a bit in very degenerate cases.