Rip out qbsptree - it didnt buy any improvement.
Modularize the internal strategy - currently QHash - and make it
possible to disable the vertex processing altogether.

Currently the qgldisplaylist tests pass.  But will need to add a test or
two for the new null condition.  Also need to add tests for the colours
thing, and for checking that vertices that are very close to each other
on adding - the solids of rotation case - do get treated as the same.

Suspect that with the QHash they won't.