Deprecate QGLVertexArray & QGLGeometry.
QArray is the new black, and QGeometryData is a collection of QArrays
that replaces QGLGeometry.  However most of the things QGLGeometry did
including being a QObject, have moved to QGLSceneNode.  This makes the
scripting to QML better since items like the palette can now be
properties on the QGLSceneNode, instead of being accessed via
the clumsy node->geometry()->palette() as before.

Also means that in the finalize step of QGLDisplayList data is posted
directly from the QArray instances inside QGeometryData, with no or
minimal copying.  QGeometryData now provides an upload() and draw()
function which do the needful things.

At this point QGLDisplayList has been ported away from QGLVertexArray
and QGLGeometry, and the tests for QGLDisplayList and QGLSection
both pass.

Still need to confirm that apps run, models load, and chase down
remaining usages of the deprecated classes.