Turn off interleaved buffers, tidy up.
Turns out the QGLVertexBuffer's interleaved mode is not working right
now and that was causing the problems.  Turning this mode off means
we're golden, although the better locality of cachelines on larger
models would mean that interleaved mode should be fixed when possible.

Also add some doco for the normals debugging - that is very handy and
worth leaving in for now - enabled by a compile time define.