Improve test coverage, fix bugs.
Add a test for the geometry building part of QGLDisplayList which really
wasn't covered.  Need to do actual stats but pretty much all of the API
is now covered.  Also add test suite for QGLOperation, and fix bugs in
QGeometryData exposed by this.  Still no test for QGeometryData, but
since QGLSection is a QGeometryData and the other tests exercise it
pretty well those tests can come in time.

Also add a warning to catch the case where the user tries to add
geometry with different field counts to the same section.  Did this
myself several times in the QGLOperation tests and it produced a hard to
track down bug.  The warning works well to save from this.