Make centroid and bb available at the scene node.
When a scene node references all of the geometry it points to its fine,
you could just call geometry()->boundingBox() - but when the node is
only pointing to part of it you have problems.

Make a nice API for getting the boundingBox and center that passes
through to the geometry() when possible, otherwise calculates the center
and bounding box for the sub section.