Add mesh modulation to 3ds plugin.
This is the first cut and needs debugging.

The N900 model seems to be flawed in that chiseled edges around the logo
on the back are not in seperate smoothing groups.  Smoothing these sharp
angles produces ugly artifacts.

The modulation function is intended to find acute angles within a
smoothing group and put the faces into different smoothing groups.

This requires building a face adjacency map and traversing it.

The same adjacency traverse can be used to detect inverted normals so
this has been added also.

Note that the GLC player program seems to display the N900 "correctly".
However the GLC player can also correct inverted normals in a test
model, so we know for certain that it is correcting at least one class
of broken model behaviour.  It needs an adjacency map to do normal
correction so it may well be doing modulation also.

Other viewer programs show the N900 with the ugly artifacts.