QML texture wrangling.
Attempt to create a demo in QML of a button object that gets its
textures updated by QML actions.

Added the demo qml file, and resources including a mesh for the button.

The current thing of setting the texture on an effect is wrong, and
doesn't work for submeshes.

Added the ability to set the url of a texture but currently this does
not seem to work either.

Suspect a rewrite of the Item3d rendering code is needed for this to
actually work.