Fix QGLView's rotation navigation functionality.
Previously, rotating the object would multiply out hundreds of tiny
micro-quaternions for each mouse twitch.  This change waits until the
mouse is released before committing the quaternion corresponding to the
entire movement.  This makes it easier to move the mouse around, return
it to the original location, and get the effect of not rotating at all.