Make geometry() be a value.
A nasty memory leak bug was sitting in the existing code because
multiple QGLSceneNode objects could refer to the same QGeometryData
object, but no system was in place to reference count the pointers, or
to have any one be a parent.  QGeometryData is not a QObject and should
not be - anyway making it one would not really fix this.

QGeometryData was already a copy-on-write, implicitly shared class, so
doing the reference counting came for free if the
QGLSceneNode::geometry() and setGeometry() methods dealt with
QGeometryData values instead of pointers.

This change was relatively simple, but lead to some annoying bugs in the
QGLDisplayList::finalize() method which puts all the sections that have
the same geometry types into one QGeometryData - these bugs are now also
fixed.

Added an operator== method to QGeometryData so that the unit tests and
other functions could use it - these call thru to operator== on the
QArray objects underneath if necessary.