Use QGLView for modelviewer.
QGLView is now able to do everything that is needed for the model
viewer, and should be used instead of reimplementing all the navigataion
functions.  Especially so since Qt3D apps should generally work-alike.

So, switch to using QGLView.  Also fix the floor, which wasn't
displaying properly.

Still needs lots of work including reading out the new orientation, zoom
and so forth from the QGLView's camera.