Skybox is now working really well.
For best results use a nice sky box texture like the ones at
http://www.hazelwhorley.com/textures.html or just pull the
research/qt3d-examples repo if you have access, then pass:
--skybox /path/to/qt3d-examples/demos/photobrowser3d/sky

The clamp mode on the textures is important, as is choosing the right
effects mode - relying on the "baked in" lighting that comes with the
textures is pretty essential to avoid the "I'm in a box" feeling.

Also needed is the adjustments to the camera to avoid bizarre
perspective issues.

Next steps - actually read in some photos, create lots of cubes
and plaster photos on, then have panning along the photo cubes
as a helix: around and along combined.