Fix bugs, improve threading.
The image loader now assigns a texture image tuned to the size of the
image being loaded.  Images with a max dimension up to 1024 get loaded
in and power-of-two sized image without scaling.  Above that they get
scaled.  In all cases they are maxspect fitted into the frame.  This
code was there before but a bug in the rect sizing was causing problems.

Also intervened in the image loading process by reading the file off
storage into memory before decoding.  This allows yielding the thread
during the read loop.  Also yield in a couple of other places.  Hope is
that this will improve things for the embedded case.