Quick basic apps with "Q" or Esc.
Current demo and example apps are very hard to quit when run full-screen
on the N900, and other devices.

Add a switch for the Q and Esc keys so they can be used to easily
exit demo and example apps.  It only works when the QGLView is the top
level widget (parentWidget() == 0).