Make build debugging accessible via env var.
When a scene is being built with QGLDisplayList, previously you could
compile in warnings for when empty geometry nodes are being generated.
These are nodes that have no children and no geometry - they are usually
a sign of a bug, in a model, or algorithm.

To turn this on took a recompile with a #define uncommented.

Make this accessible via an environment var - this makes it easier when
debugging stuff from the command line.  It can also be added to
QtCreator.  Similar idea to the env var for debugging plugins.