Make viewport lights eye-relative
Lights that are set at the viewport level are now eye-relative
and won't be affected by the modelview transform.  This has the
effect of putting the default main scene light at or behind the
camera position.  To get the previous behaviour, put the Light
object on the top-level Item3d.