QTBUG-3456 fix
This bad loop was somehow working because most models that might need
correction have just the one smoothing group because they're generated
from CAD-CAM systems.  On Linux the fail to initialize was making the
bug manifest where it was masked on Mac as it was always 0.

The larger issue of not saving values for bad models is much trickier.
You don't really know what is a "bad" model - is it taking a long time
to load or has it hung?  The only real solution is to provide a loading
dialog with a cancel button, doing the load in a background thread.

Have started coding this up and it works for many models, but not
those with textures - probably an issue with QGLContexts across
threads.  Therefore getting this fix in now, will try for the bigger
fix later.