Another interim commit.
Compiles and runs, docs build.  Tests are breaky tho'.  Have done a pass
at refactoring the stuff with common normal which was horrible, and was
causing the test breakage.  Common normals are gone now, and there is
more complexity in the addTriangles() and other methods, but it is
better code.