Update glViewport() in painter, not surface
Updating the glViewport() in the surface caused too many changes
when pushing and popping surfaces.  We move it into the painter
to be done during update() so that glViewport() will only be
changed when a drawing operation is about to be done.

Also make the surfaces return the GL-style viewport with origin
at bottom-left, because it avoids two virtual method calls to
the surface to calculate the arguments for glViewport().