Refactor and bug fix scene traversal.
Was getting crashes in a few apps still.  Time to get rid of
trouble-some memory error-prone algorithm.

Original approach was to allow custom traversal orders by sub-classing
the QGLRenderOrder.  Trouble there was since polymorphism meant having a
factory to create the custom instances, memory management was a mess.

Switched to having a seperate comparator class with override-able
isEqual and isLessThan functions, and removing the factory.  Not only
have memory headaches gone away it should be a lot faster to since no
heap allocations need to be done.  Only a single instance of the
comparator needs to be created.