Fix shader-based lighting calculations
Under certain circumstances, when the material's shininess was zero
and the angle between the normal and the half vector was 90 degrees,
the shader would calculate pow(0, 0), which is NaN.  Strange artifacts
in the specular color calculation would occur.  Modify the code to
assume that pow(x, 0) == 1 for all values of x, even x == 0.