Actually drawing something and not crashing.
The thread shutdown process now seems to be working without crashing.

Also finally hashed out the geometry and thumbnail node issues mostly so
the images are getting sent to the node, and its paining something - a
slice of a red area which looks like an atlas problem.

But its actually displaying the frames now, or trying to.  So close...