Thumbnail effect now working.
Sigh - a ; after the closing brace stops the shader from compiling.

Also made it possible to get a version with test images so that the
threading issues don't confound debugging efforts in the UI.

Refactored the thumbnail node into two nodes - the close up node is now
a seperate QGLSceneNode.  With the change to an explicitly shared
QGeometryData this is much more elegant.