Remove the 3DxWare mouse plugin
The manufacturer does not recommend the use of the 3DxWare "driver".
Instead the operating system USB HID API's should be used, which we
do via /dev/input under Linux.  The X11 protocol also doesn't give
the application much control over filtering and sensitivity so it
will never be as good as directly accessing the device.