Rename QGLFunctions to QOpenGLFunctions
There is now a QGLFunctions in Qt 4.8, but not 4.7.  This will
potentially create a binary-compatibility problem in the future
where an application compiled against a version of Qt/3D built
with Qt 4.7 won't work with a version of Qt/3D built with Qt 4.8.

Since QGLFunctions is independently useful in Qt 4.8, rename
the Qt/3D copy and then use it for all versions of Qt.