Now seems to be working nicely for a demo.
Crashy issue there for a while was due to the attempt to run the
setThumbnailed() function in the background thread.  This calls into the
global QAtlas and thus either you get concurrency issues with accessing
the image data there, or you have to lock in the GUI thread.

Result is pretty much wear the hit in the GUI thread.