Simplify the QGLAbstractScene object interface
Scene node enums such as Mesh and Main were no longer relevant,
and we had lost the ability to have arbitrary objects like
cameras and lights loaded from an external file.  Refactor the
scene API to return a QList of QObject's, which can be anything:
QGLSceneNode, QGLCamera, QGLLightParameters, etc.  Add a special
helper called mainNode() for the common case of the main mesh.