Various optimizations.
Replace the QImage based texture for the atlas with an FBO.  Big speed
improvement for the thumbnailing here.

Reduce the size of the sample images drastically, for quicker compiles
and also use these when the real pictures directory is either not
supplied or not found.

Fix a bug that was stopping the exit of the app when lots of loads were
queued up.

Remove debug.