New class: QGLAttributeSet
QGLAttributeSet is intended for "set of QGL::VertexAttribute" usages,
indicating the fields on a piece of geometry or the currently set
attributes on a QGLPainter.

It is now a good idea to call QGLPainter::clearAttributes() before
calls to setVertexBundle() or setVertexAttribute() to clear out the
"these attributes are currently available" state in the painter.