Simplify background handling for Viewport element
Rectangle and Image elements can be used to define static
backdrops, so there is no need for explicit support for
"backdrop" and "backgroundColor".  The main useful
"backgroundColor" is "transparent", to overpaint on top of
existing scene elements.  Make this the Viewport default,
and only clear to black if the Viewport does not have a parent.