Add some basic tests, fix minor obj loading issues.
Following a report of issues loading larger obj files, write some tests.

Issues mostly turned out to be with bad assumptions in the obj file,
where the use of unsupported texture coordinates was forcing duplicates
of vertices.

Tests exposed some inconsistencies - obj loader puts the node names into
the leaf nodes, this means that if a node is named and then a change
from say textured to untextured triggers a new section, then the name
will not apply to the new section.  Also sections could get created when
not needed.  Fixed these issues.