Now works for lit materials.
Most of the time spent getting to this point from the last commit was
figuring out why the models always blew out to white.  The issue may
have been due to crappy models from the blender exporter, but the
glc_player program managed to view them just fine.

Turned out to be that the default value for shininess is 0, and if the
specular color is white, and the shininess goes to its default the
specular just dominates everywhere.

To get around this if the specular color is set to something (other than
its default of black) then, set a default for the shininess which is
non-zero, to avoid this problem.  glc_player must be doing something
like this to work around bad models.