Try adding a button to center the view.
Some models are so crazy when you load them, that you can't see what is
going on - the model is very large, or moved away from the origin.

The button tries to center the view on the center of the model, and to
pull the camera back so that he model fits in the view window.