Fix bugs demonstrated by new unit test.
Still need to go thru these tests and do the math from first principles
to confirm that the collective transforms are actually doing what we
want.  At present they are correctly doing something reasonable, and
doing it consistently.

As a result the offset duck.dae model that is used by assimp, and which
caused lots of problems with being falsely culled and view sizing not
working - now works.